move Inf related classes to ota_metadata.py, add support for new hardlink group scheme
